refactor(indexeddb): move custom bool serializer into serializer module
