+++ crates/matrix-sdk-indexeddb/src/crypto_store/mod.rs
@@ -1733,7 +1733,7 @@ struct GossipRequestIndexedDbObject {
     #[serde(
         default,
         skip_serializing_if = "std::ops::Not::not",
-        with = "crate::serialize_bool_for_indexeddb"
+        with = "crate::serializer::foreign::bool"
     )]
     unsent: bool,
 }
@@ -1766,7 +1766,7 @@ struct InboundGroupSessionIndexedDbObject {
     #[serde(
         default,
         skip_serializing_if = "std::ops::Not::not",
-        with = "crate::serialize_bool_for_indexeddb"
+        with = "crate::serializer::foreign::bool"
     )]
     needs_backup: bool,
