+++ crates/matrix-sdk-indexeddb/src/serializer/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod foreign;
+
 pub mod indexed_type;
 pub use indexed_type::{
     range::IndexedKeyRange,
